requirements.txt
flask==2.0.2
WTForms==3.0.1
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
python-decouple==3.5
gunicorn==20.1.0
jinja2==3.0.3
flask-restx==0.5.1
